matrix-song-bot/requirements.txt
psycopg2-binary
matrix-nio[e2e]
python-dotenv